% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mod.R
\name{convert_all}
\alias{convert_all}
\title{Convert difficulty base on mods
Base on: https://osu.ppy.sh/help/wiki/Game_Modifiers}
\usage{
convert_all(beatmap_id, enable_mod, key, method = "api", df = NA)
}
\arguments{
\item{beatmap_id}{beatmap id}

\item{enable_mod}{Mod chosen, note this will take in numeric version of mod}

\item{key}{API key, used to get betamap info}

\item{method}{Which method we're using, could be `api` (Uses api) or `local` (R dataframe)}

\item{df}{beatmap_df}
}
\description{
Convert difficulty base on mods
Base on: https://osu.ppy.sh/help/wiki/Game_Modifiers
}
