% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ar.R
\name{convert_ar}
\alias{convert_ar}
\title{Calculate AR base on mod
This function calculate new AR base on mod}
\usage{
convert_ar(ar, enable_mod)
}
\arguments{
\item{ar}{original AR}

\item{enalbe_mod}{mod applied. Note this takes in the numeric version of mod}
}
\description{
Calculate AR base on mod
This function calculate new AR base on mod
}
