% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bpm.R
\name{convert_bpm}
\alias{convert_bpm}
\title{Convert BPM base on mod}
\usage{
convert_bpm(bpm, enabled_mods)
}
\arguments{
\item{bpm}{beatmap bpm}

\item{enabled_mods}{Mods used}
}
\description{
Convert BPM base on mod
}
