% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_drain.R
\name{convert_drain}
\alias{convert_drain}
\title{Convert drain base on mod}
\usage{
convert_drain(diff_drain, enabled_mods)
}
\arguments{
\item{diff_drain}{Drain difficulty}

\item{enabled_mods}{Mods used}
}
\description{
Convert drain base on mod
}
