% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_hit_length.R
\name{convert_hit_length}
\alias{convert_hit_length}
\title{Convert hit length base on mods}
\usage{
convert_hit_length(hit_length, enabled_mods)
}
\arguments{
\item{hit_length}{Beatmap length}

\item{enabled_mods}{Mod used}
}
\description{
Convert hit length base on mods
}
