% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mod.R
\name{convert_overall}
\alias{convert_overall}
\title{Convert overall difficulrt}
\usage{
convert_overall(diff_overall, enabled_mods)
}
\arguments{
\item{diff_overall}{Beatmap overall difficulty}

\item{enabled_mods}{Mod used}
}
\description{
Convert overall difficulrt
}
