% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mod.R
\name{convert_size}
\alias{convert_size}
\title{Convert size base on mods}
\usage{
convert_size(diff_size, enabled_mods)
}
\arguments{
\item{diff_size}{Beatmap difficulty}

\item{enabled_mods}{mods used}
}
\description{
Convert size base on mods
}
