% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_mod.R
\name{detect_mod}
\alias{detect_mod}
\title{Convert bitwise to character}
\usage{
detect_mod(value)
}
\arguments{
\item{value}{bitwise value, see https://github.com/ppy/osu-api/wiki#mods}
}
\description{
Convert bitwise to character
}
