% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beatmap.R
\name{get_beatmap}
\alias{get_beatmap}
\title{Get Beatmap info}
\usage{
get_beatmap(beatmap_id, api_key = NULL)
}
\arguments{
\item{beatmap_id}{Beatmap id}

\item{api_key}{osu! api key}
}
\value{
tibble()
}
\description{
Get Beatmap info
}
