% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player.R
\name{get_player}
\alias{get_player}
\title{Get player info}
\usage{
get_player(user, key, type = "string")
}
\arguments{
\item{user}{Username or user_id. If using user_id, need to set
\code{type} to \code{id}}

\item{key}{osu! API key}

\item{type}{Specify if user is \code{string} or \code{id}}
}
\description{
\code{get_player} return a tibble dataframe with player's info
}
