% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_convert.R
\name{mod_convert}
\alias{mod_convert}
\title{Convert difficulty base on mods
Base on: https://osu.ppy.sh/help/wiki/Game_Modifiers}
\usage{
mod_convert(beatmap_id, enable_mod, key)
}
\arguments{
\item{beatmap_id}{beatmap id}

\item{enable_mod}{Mod chosen, note this will take in numeric version of mod}

\item{key}{API key, used to get betamap info}
}
\description{
Convert difficulty base on mods
Base on: https://osu.ppy.sh/help/wiki/Game_Modifiers
}
