% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_detect.R
\name{mod_detect}
\alias{mod_detect}
\title{Convert mod value to list of mods}
\usage{
mod_detect(value)
}
\arguments{
\item{value}{enable_mod value}
}
\description{
\code{mod_detcts} converts enable_mod value to a list of character for selected mods
For full mod list see https://github.com/ppy/osu-api/wiki#mods?t
}
