% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_total_length.R
\name{osu_convert_total_length}
\alias{osu_convert_total_length}
\title{Convert total length base on mod}
\usage{
osu_convert_total_length(total_length, enabled_mods)
}
\arguments{
\item{total_length}{total length in beatmap}

\item{enabled_mods}{mod used}
}
\description{
Convert total length base on mod
}
