% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_detect.R
\name{osu_detect_mod}
\alias{osu_detect_mod}
\title{Convert mod value to list of mods}
\usage{
osu_detect_mod(value)
}
\arguments{
\item{value}{enable_mod value}
}
\description{
\code{mod_detcts} converts enable_mod value to a list of character for selected mods
For full mod list see https://github.com/ppy/osu-api/wiki#mods?t
}
