% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_df.R
\name{osu_get_player_full}
\alias{osu_get_player_full}
\title{Get player pp and bm info}
\usage{
osu_get_player_full(usr, key = key)
}
\arguments{
\item{usr}{user name}

\item{key}{API key}
}
\description{
Get player pp and bm info
}
