% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_pp.R
\name{osu_get_player_pp}
\alias{osu_get_player_pp}
\title{Get player pp info}
\usage{
osu_get_player_pp(user, type = "string", api_key = key, limit = 100)
}
\arguments{
\item{user}{Username or user_id. If using user_id, need to set
\code{type} to \code{id}}

\item{type}{Specify if user is \code{string} or \code{id}}

\item{api_key}{api key}

\item{limit}{Amount of results. Default to max 100}
}
\description{
\code{get_player_pp} return a tibble dataframe with player's top 100 plays
}
