% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_acc.R
\name{osu_map_acc}
\alias{osu_map_acc}
\title{Calculate map accuracy}
\usage{
osu_map_acc(count50, count100, count300)
}
\arguments{
\item{count50}{Total number of 50s}

\item{count100}{Total number of 100s}

\item{count300}{Total number of 300s}
}
\description{
\code{map_acc} Calcualtes the map accuracy base on number of 300, 100 and 50s.
Calculation base on https://osu.ppy.sh/help/wiki/Accuracy
}
