% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mod.R
\name{convert_ar}
\alias{convert_ar}
\title{Calculate AR base on mod
This function calculate new AR base on mod}
\usage{
convert_ar(ar, enable_mod)
}
\arguments{
\item{ar}{original AR}

\item{enable_mod}{mod used}
}
\description{
Calculate AR base on mod
This function calculate new AR base on mod
}
