% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mod.R
\name{convert_bpm}
\alias{convert_bpm}
\title{Convert BPM}
\usage{
convert_bpm(bpm, enabled_mods)
}
\arguments{
\item{bpm}{Beatmap's original BPM}

\item{enabled_mods}{Mod(s) used}
}
\value{
Converted BPM
}
\description{
Convert BPM
}
