% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beatmap.R
\name{get_beatmap}
\alias{get_beatmap}
\title{Get beatmap info}
\usage{
get_beatmap(beatmap_id, key)
}
\arguments{
\item{beatmap_id}{Beatmap Id}

\item{key}{osu! API key}
}
\description{
\code{get_beatmap} return a tibble dataframe with the appropriate class
}
