% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_pp.R
\name{get_player_pp}
\alias{get_player_pp}
\title{Get player pp info}
\usage{
get_player_pp(user, key, type = "string", limit = 100)
}
\arguments{
\item{user}{Username or user_id. If using user_id, need to set
\code{type} to \code{id}}

\item{key}{osu! API key}

\item{type}{Specify if user is \code{string} or \code{id}}

\item{limit}{Amount of results. Default to max 100}
}
\description{
\code{get_player_pp} return a tibble dataframe with player's top 100 plays
}
