% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_convert.R
\name{mod_convert}
\alias{mod_convert}
\title{Convert mod value to list of mods}
\usage{
mod_convert(value)
}
\arguments{
\item{value}{enable_mod value}
}
\description{
\code{mod_convert} converts enable_mod value to a list of character for selected mods
}
